refactor: (#798) GetModelStudentsUseCase 및 컨트롤러 students 도메인 이동
